refactor: 패키지 분리 작업 진행
