--- src/main/java/Application.java
@@ -1,8 +1,0 @@
-import controller.LottoController;
-
-public class Application {
-
-	public static void main(String[] args) {
-		new LottoController().run();
-	}
-}
