--- src/main/java/domain/TicketGenerator.java
@@ -1,8 +1,0 @@
-package domain;
-
-import java.util.List;
-
-public interface TicketGenerator {
-
-	List<Integer> generate();
-}
