+++ src/main/java/lotto/domain/ticket/generator/TicketGenerator.java
@@ -1,0 +1,8 @@
+package lotto.domain.ticket.generator;
+
+import java.util.List;
+
+public interface TicketGenerator {
+
+	List<Integer> generate();
+}
