+++ src/test/java/lotto/domain/ticket/CustomTicketGenerator.java
@@ -1,0 +1,19 @@
+package lotto.domain.ticket;
+
+import lotto.domain.ticket.generator.TicketGenerator;
+import java.util.Iterator;
+import java.util.List;
+
+class CustomTicketGenerator implements TicketGenerator {
+
+	private Iterator<List<Integer>> numbers;
+
+	public void initNumbers(List<List<Integer>> numbers) {
+		this.numbers = numbers.iterator();
+	}
+
+	@Override
+	public List<Integer> generate() {
+		return this.numbers.next();
+	}
+}
